+++ src/migrations/m161124_163823_create_email_template_table.php
@@ -4,9 +4,10 @@ use yii\db\Migration;
 /**
  * Handles the creation of table `email_template`.
  *
- * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
  * @link https://github.com/black-lamp/yii2-email-templates
- * @license https://opensource.org/licenses/GPL-3.0 GNU Public License
+ * @license GNU Public License
+ * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ * @copyright Copyright (c) Vladimir Kuprienko
  */
 class m161124_163823_create_email_template_table extends Migration
 {